Added the AutoDeleteLogs property to the FTP HTTP Log Provider settings.

+++ Hyprsoft.Cloud.Utilities/HttpLogs/FtpHttpLogProvider.cs
@@ -10,20 +10,20 @@ namespace Hyprsoft.Cloud.Utilities.HttpLogs
 {
     public class FtpHttpLogProvider : LocalHttpLogProvider
     {
-        #region Fields
-
-        private readonly FtpHttpLogProviderSettings _settings;
-
-        #endregion
-
         #region Constructors
 
         public FtpHttpLogProvider(FtpHttpLogProviderSettings settings) : base()
         {
-            _settings = settings ?? throw new ArgumentNullException(nameof(settings));
-            if (!_settings.IsValid())
+            Settings = settings ?? throw new ArgumentNullException(nameof(settings));
+            if (!Settings.IsValid())
                 throw new ArgumentOutOfRangeException("FTP HTTP log provider settings are missing or invalid.");
         }
+
+        #endregion
+
+        #region Properties
+
+        public FtpHttpLogProviderSettings Settings { get; }
 
         #endregion
 
@@ -33,21 +33,35 @@ namespace Hyprsoft.Cloud.Utilities.HttpLogs
         {
             // This FtpClient doesn't work on on Linux :-(
             // https://github.com/robinrodricks/FluentFTP/issues/347
-            using (var client = new FtpClient(_settings.Host, _settings.Username, _settings.Password)
+            using (var client = new FtpClient(Settings.Host, Settings.Username, Settings.Password)
             {
                 EncryptionMode = FtpEncryptionMode.Explicit
             })
             {
-                Logger?.LogInformation($"Connecting to FTP host '{_settings.Host}'.");
+                Logger?.LogInformation($"Connecting to FTP host '{Settings.Host}'.");
                 await client.ConnectAsync(cancellationToken).ConfigureAwait(false);
                 Logger?.LogInformation("Connected.  Getting remote files list.");
-                foreach (var file in await client.GetListingAsync(_settings.LogsFolder, cancellationToken).ConfigureAwait(false))
+                foreach (var file in await client.GetListingAsync(Settings.LogsFolder, cancellationToken).ConfigureAwait(false))
                 {
+                    if (String.Compare(Path.GetExtension(file.FullName), ".log", true) != 0)
+                        continue;
+
                     Logger?.LogInformation($"Downloading remote file '{file.FullName}'.");
                     await client.DownloadFileAsync(Path.Combine(LocalLogsFolder, file.Name), file.FullName, FtpLocalExists.Overwrite, FtpVerify.None, null, cancellationToken).ConfigureAwait(false);
-                    Logger?.LogInformation($"Deleting remote file '{file.FullName}'.");
-                    // Once our log is downloaded don't allow our delete to be cancelled.
-                    await client.DeleteFileAsync(file.FullName).ConfigureAwait(false);
+                    // Once our log is downloaded don't allow our delete or rename to be cancelled.
+                    if (Settings.AutoDeleteLogs)
+                    {
+                        Logger?.LogInformation($"Deleting remote file '{file.FullName}'.");
+                        await client.DeleteFileAsync(file.FullName).ConfigureAwait(false);
+                    }
+                    else
+                    {
+                        var dest = $"{Settings.LogsFolder}" +
+                            $"{(Settings.LogsFolder.EndsWith("/") ? String.Empty : "/")}" +
+                            $"{Path.GetFileNameWithoutExtension(file.Name)}-{Guid.NewGuid().ToString("N")}.bak";
+                        Logger?.LogInformation($"Backing up remote file '{file.FullName}' to '{dest}'.");
+                        await client.RenameAsync(file.FullName, dest).ConfigureAwait(false);
+                    }
                     await client.DisconnectAsync().ConfigureAwait(false);
                 }   // for each file
             }   // using ftp client
